creation of 3 mehtods and unit tests
creation of 3 mehtods

- GetOutlier
This method receive the array entirely comprised of odd integers or entirely comprised of even integers except for a single integer N.
This method returns the N.

- GetIntegersFromList
The method receive a list of non-negative integers and strings and returns a new list with the strings filtered out.

- FormatWords
The method formats the words into a single comma separated value. The last word is separated by the word 'and'.

Also, I created the unit tests for the 3 methods.